refactor: adjust SchemaGenerator parameters and centralize prompts
